added backward compatibility
