+++ shell/directives/clean-html.js
@@ -1,0 +1,15 @@
+import { purifyHTML } from '@shell/plugins/clean-html';
+
+const cleanHtmlDirective = {
+  inserted(el, binding) {
+    el.innerHTML = purifyHTML(binding.value);
+  },
+  componentUpdated(el, binding) {
+    el.innerHTML = purifyHTML(binding.value);
+  },
+  unbind(el) {
+    el.innerHTML = '';
+  }
+};
+
+export default cleanHtmlDirective;
